added another slide to the slides
added a new file to the _posts folder for creating a new slide in the slideshow. 
The course could be updated with
suggestion```
create the branch my-slide during when committing the file. the Github UI and UX has been updated.
```
tested by creating another branch > switching to the new branch > creating a new file > commit > only option: create new branch
the branch name was not important as  creating any branch with any file made the learning assistant think we completed the task